checking delete master data

+++ application/models/Mbank_model.php
@@ -21,6 +21,12 @@ class Mbank_model extends CI_Model {
 
   public function delete($id)
   {
+    $this->db->where('id_bank', $id);
+    $query = $this->db->get('tb_detail_penjualan');
+
+    if ($query->num_rows() > 0) {
+      return array('success' => false, 'message' => 'Data bank dengan nomor rekening "'.$id.'" tidak bisa dihapus, karena sudah terdaftar dipenjualan');
+    }   
     $success = $this->db->delete('tb_bank', array("id_bank" => $id));
     $message = $success ? 'Data berhasil dihapus' : 'Gagal dihapus';
     return array('success' => $success, 'message' => $message);
