+++ application/models/Mcabang_model.php
@@ -51,6 +51,12 @@ class Mcabang_model extends CI_Model {
 
   public function delete($id)
   {
+    $this->db->where('id_toko', $id);
+    $query = $this->db->get('tb_brg_keluar');
+
+    if ($query->num_rows() > 0) {
+        return array('success' => false, 'message' => 'Data cabang dengan id "'.$id.'" tidak bisa dihapus');
+    }
     $success = $this->db->delete('tb_toko', array("id_toko" => $id));
     $message = $success ? 'Data berhasil dihapus' : 'Gagal dihapus';
     return array('success' => $success, 'message' => $message);
